fix unspecified dynamic client watch timeout
Currently, when a dynamic client watcher is used, timeout_seconds is
always passed to watcher.stream() from DynamicClient.watch(). Its value
is None when the timeout argument was not passed by the user.

The current check done on TimeoutError in the Watch.watch() method when
awaiting from self.resp.content.readline() prevents to reopen the
request in this case because it behaves differently if timeout_seconds
was set to None or if timeout_seconds was not passed to the function.

Do not pass the timeout_seconds argument in DynamicClient.watch() if
timeout was not specified by the user.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>